Failed attempt to add TIMPI::Attributes specializations
It's really not easy or it may even be impossible to implement
verification for `MetaPhysicL` types. This is because you would have to
write specializations for `TIMPI::Communicator::min` and `max` for the
`MetaPhysicL` types and you can only fully specialize function
templates. So we would just have to do a million full template
specialization declarations, and the reward is just not worth the work.
Instead I'm just going to call `verify` with `MetaPhysicL::raw_value`